Extract tests of Prelude
